Refactor boot image factory
Make sure the root directory of the base image is always
accessible by any boot image type for consistency. In
addition introduce a post_init method which explicitly
setup the boot image root directory as needed for the
selected boot image type and document it as such
